pyproject.toml
[tool.poetry]
name = "xls_handler"
version = "0.1.0"
description = "xls_handler"
authors = ["Dobro"]
packages = [
  { include = "xls_handler" },
]

[tool.poetry.dependencies]
python = "^3.8"
openpyxl = "^3.0.9"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.15.3"
autopep8 = "^1.5.7"
isort = "^5.9.3"
pytest-cov = "^3.0.0"
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
xls-handler = "xls_handler.xls_handler:main"